+++ examples/angular_counter/web/main.dart
@@ -1,22 +1,22 @@
 import 'package:angular/angular.dart';
-import 'package:true_bloc/bloc.dart';
 import 'package:angular_counter/app_component.template.dart' as ng;
+import 'package:true_bloc/bloc.dart';
 
 class SimpleBlocObserver extends BlocObserver {
   @override
-  void onEvent(Bloc bloc, Object? event) {
+  void onEvent(IBloc bloc, Object? event) {
     print(event);
     super.onEvent(bloc, event);
   }
 
   @override
-  void onTransition(Bloc bloc, Transition transition) {
+  void onTransition(IBloc bloc, Transition transition) {
     print(transition);
     super.onTransition(bloc, transition);
   }
 
   @override
-  void onError(BlocBase bloc, Object error, StackTrace stackTrace) {
+  void onError(IBlocBase bloc, Object error, StackTrace stackTrace) {
     print(error);
     super.onError(bloc, error, stackTrace);
   }