+++ examples/flutter_firebase_login/lib/app/bloc_observer.dart
@@ -3,25 +3,25 @@ import 'package:true_bloc/bloc.dart';
 
 class AppBlocObserver extends BlocObserver {
   @override
-  void onEvent(Bloc bloc, Object? event) {
+  void onEvent(IBloc bloc, Object? event) {
     super.onEvent(bloc, event);
     print(event);
   }
 
   @override
-  void onError(BlocBase bloc, Object error, StackTrace stackTrace) {
+  void onError(IBlocBase bloc, Object error, StackTrace stackTrace) {
     print(error);
     super.onError(bloc, error, stackTrace);
   }
 
   @override
-  void onChange(BlocBase bloc, Change change) {
+  void onChange(IBlocBase bloc, Change change) {
     super.onChange(bloc, change);
     print(change);
   }
 
   @override
-  void onTransition(Bloc bloc, Transition transition) {
+  void onTransition(IBloc bloc, Transition transition) {
     super.onTransition(bloc, transition);
     print(transition);
   }