+++ examples/flutter_infinite_list/lib/simple_bloc_observer.dart
@@ -2,13 +2,13 @@ import 'package:true_bloc/bloc.dart';
 
 class SimpleBlocObserver extends BlocObserver {
   @override
-  void onTransition(Bloc bloc, Transition transition) {
+  void onTransition(IBloc bloc, Transition transition) {
     super.onTransition(bloc, transition);
     print(transition);
   }
 
   @override
-  void onError(BlocBase bloc, Object error, StackTrace stackTrace) {
+  void onError(IBlocBase bloc, Object error, StackTrace stackTrace) {
     print(error);
     super.onError(bloc, error, stackTrace);
   }