+++ examples/flutter_shopping_cart/lib/simple_bloc_observer.dart
@@ -4,19 +4,19 @@ import 'package:true_bloc/bloc.dart';
 
 class SimpleBlocObserver extends BlocObserver {
   @override
-  void onEvent(Bloc bloc, Object? event) {
+  void onEvent(IBloc bloc, Object? event) {
     super.onEvent(bloc, event);
     log('${bloc.runtimeType} $event');
   }
 
   @override
-  void onError(BlocBase bloc, Object error, StackTrace stackTrace) {
+  void onError(IBlocBase bloc, Object error, StackTrace stackTrace) {
     log('${bloc.runtimeType} $error');
     super.onError(bloc, error, stackTrace);
   }
 
   @override
-  void onTransition(Bloc bloc, Transition transition) {
+  void onTransition(IBloc bloc, Transition transition) {
     super.onTransition(bloc, transition);
     log('$transition');
   }