+++ examples/flutter_weather/lib/weather_bloc_observer.dart
@@ -2,25 +2,25 @@ import 'package:true_bloc/bloc.dart';
 
 class WeatherBlocObserver extends BlocObserver {
   @override
-  void onEvent(Bloc bloc, Object? event) {
+  void onEvent(IBloc bloc, Object? event) {
     super.onEvent(bloc, event);
     print('onEvent $event');
   }
 
   @override
-  void onChange(BlocBase bloc, Change change) {
+  void onChange(IBlocBase bloc, Change change) {
     super.onChange(bloc, change);
     print('onChange $change');
   }
 
   @override
-  void onTransition(Bloc bloc, Transition transition) {
+  void onTransition(IBloc bloc, Transition transition) {
     super.onTransition(bloc, transition);
     print('onTransition $transition');
   }
 
   @override
-  void onError(BlocBase bloc, Object error, StackTrace stackTrace) {
+  void onError(IBlocBase bloc, Object error, StackTrace stackTrace) {
     print('onError $error');
     super.onError(bloc, error, stackTrace);
   }