+++ packages/replay_true_bloc/example/lib/simple_bloc_observer.dart
@@ -3,25 +3,25 @@ import 'package:true_bloc/bloc.dart';
 /// Simple [BlocObserver] which just prints to the debug console.
 class SimpleBlocObserver extends BlocObserver {
   @override
-  void onEvent(Bloc bloc, Object? event) {
+  void onEvent(IBloc bloc, Object? event) {
     super.onEvent(bloc, event);
     print('${bloc.runtimeType} $event');
   }
 
   @override
-  void onError(BlocBase bloc, Object error, StackTrace stackTrace) {
+  void onError(IBlocBase bloc, Object error, StackTrace stackTrace) {
     print('${bloc.runtimeType} $error $stackTrace');
     super.onError(bloc, error, stackTrace);
   }
 
   @override
-  void onChange(BlocBase bloc, Change change) {
+  void onChange(IBlocBase bloc, Change change) {
     super.onChange(bloc, change);
     print('${bloc.runtimeType} $change');
   }
 
   @override
-  void onTransition(Bloc bloc, Transition transition) {
+  void onTransition(IBloc bloc, Transition transition) {
     super.onTransition(bloc, transition);
     print('${bloc.runtimeType} $transition');
   }