+++ packages/true_bloc/lib/src/bloc.dart
@@ -1,72 +1,37 @@
 import 'dart:async';
 
 import 'package:meta/meta.dart';
-import 'package:true_bloc/bloc.dart';
+
+part 'bloc_base.dart';
+part 'bloc_observer.dart';
+part 'bloc_unhandled_error_exception.dart';
+part 'cubit.dart';
+part 'emitter_mixin.dart';
+part 'transition.dart';
 
 /// Signature for a mapper function which takes an [Event] as input
 /// and outputs a [Stream] of [Transition] objects.
-typedef TransitionFunction<Event, State> = Stream<Transition<Event, State>>
-    Function(Event);
+typedef TransitionFunction<Event extends Object?, State extends Object?>
+    = Stream<Transition<Event, State>> Function(Event);
 
-/// {@template bloc_unhandled_error_exception}
-/// Exception thrown when an unhandled error occurs within a bloc.
-///
-/// _Note: thrown in debug mode only_
-/// {@endtemplate}
-class BlocUnhandledErrorException implements Exception {
-  /// {@macro bloc_unhandled_error_exception}
-  BlocUnhandledErrorException(
-    this.bloc,
-    this.error, [
-    this.stackTrace = StackTrace.empty,
-  ]);
-
-  /// The bloc in which the unhandled error occurred.
-  final BlocBase bloc;
-
-  /// The unhandled [error] object.
-  final Object error;
-
-  /// Stack trace which accompanied the error.
-  /// May be [StackTrace.empty] if no stack trace was provided.
-  final StackTrace stackTrace;
-
-  @override
-  String toString() {
-    return 'Unhandled error $error occurred in $bloc.\n'
-        '$stackTrace';
-  }
-}
-
-/// {@template bloc}
+/// {@template i_bloc}
+/// An interface for [Bloc].
 /// Takes a `Stream` of `Events` as input
 /// and transforms them into a `Stream` of `States` as output.
 /// {@endtemplate}
-abstract class Bloc<Event, State> extends BlocBase<State> {
-  /// {@macro bloc}
-  Bloc(State initialState) : super(initialState) {
-    _bindEventsToStates();
-  }
+abstract class IBloc<Event extends Object?, State extends Object?>
+    extends IBlocBase<State> implements StreamSink<Event> {
+  /// Pass all [eventStream] events to the [Bloc]
+  /// If [close] has already been called, any subsequent calls to [addStream]
+  /// will be ignored and will not result in any subsequent state changes.
+  @override
+  Future<void> addStream(Stream<Event> eventStream);
 
-  /// The current [BlocObserver] instance.
-  static BlocObserver observer = BlocObserver();
-
-  StreamSubscription<Transition<Event, State>>? _transitionSubscription;
-
-  final _eventController = StreamController<Event>.broadcast();
-
-  /// Notifies the [Bloc] of a new [event] which triggers [mapEventToState].
+  /// Notifies the [Bloc] of a new [event]
   /// If [close] has already been called, any subsequent calls to [add] will
   /// be ignored and will not result in any subsequent state changes.
-  void add(Event event) {
-    if (_eventController.isClosed) return;
-    try {
-      onEvent(event);
-      _eventController.add(event);
-    } catch (error, stackTrace) {
-      onError(error, stackTrace);
-    }
-  }
+  @override
+  void add(Event event);
 
   /// Called whenever an [event] is [add]ed to the [Bloc].
   /// A great spot to add logging/analytics at the individual [Bloc] level.
@@ -84,14 +49,12 @@ abstract class Bloc<Event, State> extends BlocBase<State> {
   ///
   /// See also:
   ///
-  /// * [BlocObserver.onEvent] for observing events globally.
+  /// * `IBlocObserver.onEvent` for observing events globally.
   ///
   @protected
   @mustCallSuper
-  void onEvent(Event event) {
-    // ignore: invalid_use_of_protected_member
-    observer.onEvent(this, event);
-  }
+  @visibleForOverriding
+  void onEvent(Event event);
 
   /// Transforms the [events] stream along with a [transitionFn] function into
   /// a `Stream<Transition>`.
@@ -125,29 +88,18 @@ abstract class Bloc<Event, State> extends BlocBase<State> {
   ///   );
   /// }
   /// ```
+  @visibleForOverriding
   Stream<Transition<Event, State>> transformEvents(
     Stream<Event> events,
     TransitionFunction<Event, State> transitionFn,
-  ) {
-    return events.asyncExpand(transitionFn);
-  }
-
-  /// {@template emit}
-  /// **[emit] should never be used outside of tests.**
-  ///
-  /// Updates the state of the bloc to the provided [state].
-  /// A bloc's state should only be updated by `yielding` a new `state`
-  /// from `mapEventToState` in response to an event.
-  /// {@endtemplate}
-  @protected
-  @visibleForTesting
-  @override
-  void emit(State state) => super.emit(state);
+  );
 
   /// Must be implemented when a class extends [Bloc].
   /// [mapEventToState] is called whenever an [event] is [add]ed
   /// and is responsible for converting that [event] into a new [state].
   /// [mapEventToState] can `yield` zero, one, or multiple states for an event.
+  @protected
+  @visibleForOverriding
   Stream<State> mapEventToState(Event event);
 
   /// Called whenever a [transition] occurs with the given [transition].
@@ -173,10 +125,8 @@ abstract class Bloc<Event, State> extends BlocBase<State> {
   ///
   @protected
   @mustCallSuper
-  void onTransition(Transition<Event, State> transition) {
-    // ignore: invalid_use_of_protected_member
-    Bloc.observer.onTransition(this, transition);
-  }
+  @visibleForOverriding
+  void onTransition(Transition<Event, State> transition);
 
   /// Transforms the `Stream<Transition>` into a new `Stream<Transition>`.
   /// By default [transformTransitions] returns
@@ -195,11 +145,11 @@ abstract class Bloc<Event, State> extends BlocBase<State> {
   ///   return transitions.debounceTime(Duration(seconds: 1));
   /// }
   /// ```
+  @protected
+  @visibleForOverriding
   Stream<Transition<Event, State>> transformTransitions(
     Stream<Transition<Event, State>> transitions,
-  ) {
-    return transitions;
-  }
+  );
 
   /// Closes the `event` and `state` `Streams`.
   /// This method should be called when a [Bloc] is no longer needed.
@@ -209,11 +159,107 @@ abstract class Bloc<Event, State> extends BlocBase<State> {
   /// processed, the [Bloc] will finish processing the pending `events`.
   @override
   @mustCallSuper
+  @visibleForOverriding
+  Future<void> close();
+
+  /// Return a future which is completed when the [IBloc] is finished.
+  ///
+  /// If the `IBloc` fails with an error,
+  /// perhaps in response to adding events using [add], [addStream], [addError]
+  /// or [close], the [done] future will complete with that error.
+  ///
+  /// Otherwise, the returned future will complete when either:
+  ///
+  /// * all events have been processed and the sink has been closed, or
+  /// * the sink has otherwise been stopped from handling more events
+  ///   (for example by canceling a stream subscription).
+  @override
+  Future<void> get done;
+}
+
+/// {@template bloc}
+/// Takes a `Stream` of `Events` as input
+/// and transforms them into a `Stream` of `States` as output.
+/// {@endtemplate}
+abstract class Bloc<Event, State> extends BlocBase<State>
+    implements IBloc<Event, State> {
+  /// {@macro bloc}
+  Bloc(State initialState) : super(initialState) {
+    _bindEventsToStates();
+  }
+
+  /// The current [BlocObserver] instance.
+  static BlocObserver observer = BlocObserver();
+
+  StreamSubscription<Transition<Event, State>>? _transitionSubscription;
+
+  final _eventController = StreamController<Event>.broadcast();
+
+  @override
+  Future<void> addStream(Stream<Event> eventStream) =>
+      _eventController.addStream(eventStream);
+
+  @override
+  void add(Event event) {
+    if (_eventController.isClosed) return;
+    try {
+      onEvent(event);
+      _eventController.add(event);
+    } catch (error, stackTrace) {
+      onError(error, stackTrace);
+    }
+  }
+
+  @protected
+  @mustCallSuper
+  @override
+  void onEvent(Event event) {
+    // ignore: invalid_use_of_protected_member
+    observer.onEvent(this, event);
+  }
+
+  @override
+  Stream<Transition<Event, State>> transformEvents(
+    Stream<Event> events,
+    TransitionFunction<Event, State> transitionFn,
+  ) {
+    return events.asyncExpand(transitionFn);
+  }
+
+  @protected
+  @visibleForTesting
+  @override
+  @Deprecated(
+    'Use EmitterMixin instead. Will be removed in v8.0.0',
+  )
+  void emit(State state) => super.emit(state);
+
+  @protected
+  @mustCallSuper
+  @override
+  void onTransition(Transition<Event, State> transition) {
+    // ignore: invalid_use_of_protected_member
+    Bloc.observer.onTransition(this, transition);
+  }
+
+  @override
+  Stream<Transition<Event, State>> transformTransitions(
+    Stream<Transition<Event, State>> transitions,
+  ) {
+    return transitions;
+  }
+
+  @override
+  @mustCallSuper
+  @override
   Future<void> close() async {
     await _eventController.close();
     await _transitionSubscription?.cancel();
     return super.close();
   }
+
+  @override
+  Future<void> get done => _eventController.done;
 
   void _bindEventsToStates() {
     _transitionSubscription = transformTransitions(
@@ -239,167 +285,5 @@ abstract class Bloc<Event, State> extends BlocBase<State> {
       },
       onError: onError,
     );
-  }
-}
-
-/// {@template cubit}
-/// A [Cubit] is similar to [Bloc] but has no notion of events
-/// and relies on methods to [emit] new states.
-///
-/// Every [Cubit] requires an initial state which will be the
-/// state of the [Cubit] before [emit] has been called.
-///
-/// The current state of a [Cubit] can be accessed via the [state] getter.
-///
-/// ```dart
-/// class CounterCubit extends Cubit<int> {
-///   CounterCubit() : super(0);
-///
-///   void increment() => emit(state + 1);
-/// }
-/// ```
-///
-/// {@endtemplate}
-abstract class Cubit<State> extends BlocBase<State> {
-  /// {@macro cubit}
-  Cubit(State initialState) : super(initialState);
-}
-
-/// {@template bloc_stream}
-/// An interface for the core functionality implemented by
-/// both [Bloc] and [Cubit].
-/// {@endtemplate}
-abstract class BlocBase<State> {
-  /// {@macro bloc_stream}
-  BlocBase(this._state) {
-    // ignore: invalid_use_of_protected_member
-    Bloc.observer.onCreate(this);
-  }
-
-  StreamController<State>? __stateController;
-  StreamController<State> get _stateController {
-    return __stateController ??= StreamController<State>.broadcast();
-  }
-
-  State _state;
-
-  bool _emitted = false;
-
-  /// The current [state].
-  State get state => _state;
-
-  /// The current state stream.
-  Stream<State> get stream => _stateController.stream;
-
-  /// Whether the bloc is closed.
-  ///
-  /// A bloc is considered closed once [close] is called.
-  /// Subsequent state changes cannot occur within a closed bloc.
-  bool get isClosed => _stateController.isClosed;
-
-  /// Adds a subscription to the `Stream<State>`.
-  /// Returns a [StreamSubscription] which handles events from
-  /// the `Stream<State>` using the provided [onData], [onError] and [onDone]
-  /// handlers.
-  @Deprecated(
-    'Use stream.listen instead. Will be removed in v8.0.0',
-  )
-  StreamSubscription<State> listen(
-    void Function(State)? onData, {
-    Function? onError,
-    void Function()? onDone,
-    bool? cancelOnError,
-  }) {
-    return stream.listen(
-      onData,
-      onError: onError,
-      onDone: onDone,
-      cancelOnError: cancelOnError,
-    );
-  }
-
-  /// Updates the [state] to the provided [state].
-  /// [emit] does nothing if the instance has been closed or if the
-  /// [state] being emitted is equal to the current [state].
-  ///
-  /// To allow for the possibility of notifying listeners of the initial state,
-  /// emitting a state which is equal to the initial state is allowed as long
-  /// as it is the first thing emitted by the instance.
-  void emit(State state) {
-    if (_stateController.isClosed) return;
-    if (state == _state && _emitted) return;
-    onChange(Change<State>(currentState: this.state, nextState: state));
-    _state = state;
-    _stateController.add(_state);
-    _emitted = true;
-  }
-
-  /// Called whenever a [change] occurs with the given [change].
-  /// A [change] occurs when a new `state` is emitted.
-  /// [onChange] is called before the `state` of the `cubit` is updated.
-  /// [onChange] is a great spot to add logging/analytics for a specific `cubit`.
-  ///
-  /// **Note: `super.onChange` should always be called first.**
-  /// ```dart
-  /// @override
-  /// void onChange(Change change) {
-  ///   // Always call super.onChange with the current change
-  ///   super.onChange(change);
-  ///
-  ///   // Custom onChange logic goes here
-  /// }
-  /// ```
-  ///
-  /// See also:
-  ///
-  /// * [BlocObserver] for observing [Cubit] behavior globally.
-  @mustCallSuper
-  void onChange(Change<State> change) {
-    // ignore: invalid_use_of_protected_member
-    Bloc.observer.onChange(this, change);
-  }
-
-  /// Reports an [error] which triggers [onError] with an optional [StackTrace].
-  @mustCallSuper
-  void addError(Object error, [StackTrace? stackTrace]) {
-    onError(error, stackTrace ?? StackTrace.current);
-  }
-
-  /// Called whenever an [error] occurs and notifies [BlocObserver.onError].
-  ///
-  /// In debug mode, [onError] throws a [BlocUnhandledErrorException] for
-  /// improved visibility.
-  ///
-  /// In release mode, [onError] does not throw and will instead only report
-  /// the error to [BlocObserver.onError].
-  ///
-  /// **Note: `super.onError` should always be called last.**
-  /// ```dart
-  /// @override
-  /// void onError(Object error, StackTrace stackTrace) {
-  ///   // Custom onError logic goes here
-  ///
-  ///   // Always call super.onError with the current error and stackTrace
-  ///   super.onError(error, stackTrace);
-  /// }
-  /// ```
-  @protected
-  @mustCallSuper
-  void onError(Object error, StackTrace stackTrace) {
-    // ignore: invalid_use_of_protected_member
-    Bloc.observer.onError(this, error, stackTrace);
-    assert(() {
-      throw BlocUnhandledErrorException(this, error, stackTrace);
-    }());
-  }
-
-  /// Closes the instance.
-  /// This method should be called when the instance is no longer needed.
-  /// Once [close] is called, the instance can no longer be used.
-  @mustCallSuper
-  Future<void> close() async {
-    // ignore: invalid_use_of_protected_member
-    Bloc.observer.onClose(this);
-    await _stateController.close();
   }
 }