+++ packages/true_bloc/lib/true_bloc.dart
@@ -1,3 +1,3 @@
-library bloc;
+library true_bloc;
 
 export './src/bloc.dart';